Restructure the backend module
